+++ AndroidAssignment1/app/build.gradle
@@ -35,6 +35,11 @@ android {
 }
 
 dependencies {
+
+    // navigation component
+    implementation 'androidx.navigation:navigation-fragment-ktx:2.3.5'
+    implementation 'androidx.navigation:navigation-ui-ktx:2.3.5'
+
     // 서버 연결을 위한 Retrofit2
     implementation "com.squareup.retrofit2:retrofit:2.9.0"
