+++ AndroidAssignment1/app/src/main/java/org/sopt/androidassignment1/config/ItemTouchHelperCallback.kt
@@ -1,4 +1,4 @@
-package org.sopt.androidassignment1
+package org.sopt.androidassignment1.config
 
 import androidx.recyclerview.widget.ItemTouchHelper
 import androidx.recyclerview.widget.RecyclerView
