+++ AndroidAssignment1/app/src/main/java/org/sopt/androidassignment1/config/ItemTouchHelperListener.kt
@@ -1,4 +1,4 @@
-package org.sopt.androidassignment1
+package org.sopt.androidassignment1.config
 
 interface ItemTouchHelperListener {
     fun onItemMove(from_position: Int, to_position: Int):Boolean
