+++ AndroidAssignment1/app/src/main/java/org/sopt/androidassignment1/ui/DetailActivity.kt
@@ -1,7 +1,8 @@
-package org.sopt.androidassignment1
+package org.sopt.androidassignment1.ui
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import org.sopt.androidassignment1.R
 import org.sopt.androidassignment1.databinding.ActivityDetailBinding
 
 class DetailActivity : AppCompatActivity() {
