+++ AndroidAssignment1/app/src/main/java/org/sopt/androidassignment1/ui/GithubFollowerFragment.kt
@@ -1,21 +1,21 @@
-package org.sopt.androidassignment1
+package org.sopt.androidassignment1.ui
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import org.sopt.androidassignment1.databinding.FragmentGithubFollowingBinding
+import org.sopt.androidassignment1.databinding.FragmentGithubFollowerBinding
 
-class GithubFollowingFragment : Fragment() {
-    private var _binding: FragmentGithubFollowingBinding? = null
+class GithubFollowerFragment : Fragment() {
+    private var _binding: FragmentGithubFollowerBinding? = null
     private val binding get() = _binding!!
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        _binding = FragmentGithubFollowingBinding.inflate(layoutInflater, container, false)
+        _binding = FragmentGithubFollowerBinding.inflate(layoutInflater, container, false)
 
         return binding.root
     }
