+++ AndroidAssignment1/app/src/main/java/org/sopt/androidassignment1/ui/HomeFragment1.kt
@@ -1,4 +1,4 @@
-package org.sopt.androidassignment1
+package org.sopt.androidassignment1.ui
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,8 +6,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.bumptech.glide.Glide
-import com.google.android.material.tabs.TabLayoutMediator
-import org.sopt.androidassignment1.databinding.FragmentFollowerBinding
+import org.sopt.androidassignment1.R
 import org.sopt.androidassignment1.databinding.FragmentHome1Binding
 
 /* HomeActivity -> Fragment <Profile> */
