+++ AndroidAssignment1/app/src/main/java/org/sopt/androidassignment1/ui/HomeFragment3.kt
@@ -1,17 +1,12 @@
-package org.sopt.androidassignment1
+package org.sopt.androidassignment1.ui
 
-import android.Manifest
 import android.content.Intent
-import android.content.pm.PackageManager
 import android.os.Bundle
 import android.provider.MediaStore
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.Toast
-import androidx.core.app.ActivityCompat
-import androidx.core.content.ContextCompat
 import com.bumptech.glide.Glide
 import org.sopt.androidassignment1.databinding.FragmentHome3Binding
