+++ AndroidAssignment1/app/src/main/java/org/sopt/androidassignment1/ui/SignUpActivity.kt
@@ -1,4 +1,4 @@
-package org.sopt.androidassignment1
+package org.sopt.androidassignment1.ui
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
