+++ AndroidAssignment1/app/src/main/res/layout/activity_log_in.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="10dp"
-    tools:context=".LogInActivity">
+    tools:context=".ui.LogInActivity">
 
     <TextView
         android:id="@+id/tv_id"
