+++ AndroidAssignment1/app/src/main/res/layout/fragment_github_following.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#DFDFDF"
-    tools:context=".GithubFollowingFragment">
+    tools:context=".ui.GithubFollowingFragment">
 
     <TextView
         android:id="@+id/tv_ready_not"
