+++ AndroidAssignment1/app/src/main/res/layout/fragment_home1.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".HomeFragment1">
+    tools:context=".ui.HomeFragment1">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/cl_profile_background"
