+++ AndroidAssignment1/app/src/main/res/layout/fragment_home3.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".HomeFragment3">
+    tools:context=".ui.HomeFragment3">
 
 
     <ImageView
