+++ AndroidAssignment1/app/src/main/res/layout/fragment_repository.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".RepositoryFragment">
+    tools:context=".ui.RepositoryFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rv_repository"
